chapter-generation-service/requirements-chapters.txt
kafka-python
google-generativeai